requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-jazzmin==2.6.0
django-tinymce==3.6.1
idna==3.4
Pillow==10.1.0
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.7
